Merge pull request #41 from pkaff/db_synchronize_activity
Remove unnecessary items and lines of code and refactoring

Clean up the code to make it readable!!

Add sync toolbar menu item and functionality

Syncs the database when pressed. The database is entirely taken
from the server, deleting the local one.

Add settings icon for toolbar

Remove visibility for various action buttons on toolbar in various views

They are not necessary - for example the settings action button
while in the settings activity or the home button while already in the
main activity.

Move checking for internet connection inside the syncdatabase task

Then the check will also be performed on start of app and recipes list
will just be empty if no internet was found.